move fetch user history to profile page
