+++ src/pages/Storefront.jsx
@@ -6,6 +6,7 @@ import CartDrawer from "../components/storefront/CartDrawer.jsx";
 import { ShoppingCartIcon } from "../components/CommonComponents/icons.jsx";
 import { useDispatch, useSelector } from "react-redux";
 import axios from "axios";
+import { getAllItems } from "../store/globalReducer.js";
 
 
 function Storefront() {
@@ -15,30 +16,6 @@ function Storefront() {
     const [selectedItem, setSelectedItem] = useState(null);
 
     const dispatch = useDispatch()
-
-    const userInfo = useSelector((state) => {
-        return state.globalState.userProfile
-    })
-
-    // This is to fetch user history when shopper logs in. Since shoppers are navigated here first, it made sense to put here instead of profile page
-    useEffect(() => {
-        async function fetchUserHistory() {
-            try {
-                const response = await axios.get(`/api/history/${userInfo.userId}`);
-                const userHistory = response.data;
-
-                dispatch({
-                    type: 'SET_USER_HISTORY',
-                    payload: userHistory
-                });
-            } catch (error) {
-                console.error("Error fetching user history:", error);
-            }
-        }
-        if (userInfo.userId) {
-            fetchUserHistory();
-        }
-    }, [userInfo.userId, dispatch])
 
     // This is to fetch regular inventory for the shopping page
     useEffect(() => {
@@ -97,7 +74,6 @@ function Storefront() {
     }
 
     //TODO : add ability to update quanity in textbox + update quanity in general
-    //TODO : "Buy Now" subtracts points from user + subtract item from inventory + add item to history
 
     function addToCartDrawer() {
